refactored several functions out of server
